+++ Tortuga.Chain/Tortuga.Chain.SqlServer.OleDb/SqlServer/OleDbSqlServerMetadataCache.cs
@@ -496,7 +496,9 @@ public sealed partial class OleDbSqlServerMetadataCache
 				t.name AS Name,
 				t.object_id AS ObjectId,
 				CONVERT(BIT, 1) AS IsTable,
-				(SELECT	COUNT(*) FROM sys.triggers t2 WHERE	t2.parent_id = t.object_id) AS Triggers
+				(SELECT	COUNT(*) FROM sys.triggers t2 WHERE	t2.parent_id = t.object_id) AS Triggers,
+				CONVERT(BIT, CASE WHEN t.temporal_type = 1 THEN 1 ELSE 0 END) AS IsHistoryTable,
+				t.history_table_id
 				FROM SYS.tables t
 				INNER JOIN sys.schemas s ON t.schema_id = s.schema_id
 				WHERE s.name = ? AND t.Name = ?
@@ -508,7 +510,9 @@ public sealed partial class OleDbSqlServerMetadataCache
 				t.name AS Name,
 				t.object_id AS ObjectId,
 				CONVERT(BIT, 0) AS IsTable,
-				(SELECT	COUNT(*) FROM sys.triggers t2 WHERE	t2.parent_id = t.object_id) AS Triggers
+				(SELECT	COUNT(*) FROM sys.triggers t2 WHERE	t2.parent_id = t.object_id) AS Triggers,
+				CONVERT(BIT, 0) AS IsHistoryTable,
+				NULL AS history_table_id
 				FROM SYS.views t
 				INNER JOIN sys.schemas s ON t.schema_id = s.schema_id
 				WHERE s.name = ? AND t.Name = ?";
@@ -518,6 +522,8 @@ public sealed partial class OleDbSqlServerMetadataCache
 		int objectId;
 		bool isTable;
 		bool hasTriggers;
+		bool isHistoryTable;
+		int? historyTableObjectId;
 
 		using (var con = new OleDbConnection(m_ConnectionBuilder!.ConnectionString))
 		{
@@ -536,6 +542,8 @@ public sealed partial class OleDbSqlServerMetadataCache
 					actualName = reader.GetString("Name");
 					objectId = reader.GetInt32("ObjectId");
 					isTable = reader.GetBoolean("IsTable");
+					isHistoryTable = reader.GetBoolean("IsHistoryTable");
+					historyTableObjectId = reader.GetInt32OrNull("history_table_id");
 					hasTriggers = reader.GetInt32("Triggers") > 0;
 				}
 			}
@@ -546,7 +554,9 @@ public sealed partial class OleDbSqlServerMetadataCache
 		var actualTableName = new SqlServerObjectName(actualSchema, actualName);
 		m_ObjectIdTableMap[objectId] = actualTableName;
 
-		return new SqlServerTableOrViewMetadata<OleDbType>(this, actualTableName, isTable, columns, hasTriggers, objectId);
+		var historyTableName = historyTableObjectId.HasValue ? (SqlServerObjectName?)GetTableOrViewName(historyTableObjectId.Value) : null;
+
+		return new SqlServerTableOrViewMetadata<OleDbType>(this, actualTableName, isTable, columns, hasTriggers, objectId, historyTableName, isHistoryTable);
 	}
 
 	internal UserDefinedTableTypeMetadata<SqlServerObjectName, OleDbType>